+++ concert-service/src/main/java/com/bit/docker/concertservice/infra/SeatRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
 
-    List<Seat> findByConcertId(Long concertId);
+    List<Seat> findByConcertId(int concertId);
 }
